move: demo.domain 패키지로 이동
