+++ src/main/java/com/example/demo/domain/column/dto/RequestColumnDto.java
@@ -1,4 +1,4 @@
-package com.example.demo.column.dto;
+package com.example.demo.domain.column.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
