+++ src/test/java/com/example/demo/column/service/ColumnServiceTest.java
@@ -1,0 +1,20 @@
+package com.example.demo.column.service;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class ColumnServiceTest {
+
+    @Test
+    void createColumn() {
+    }
+
+    @Test
+    void deleteColumn() {
+    }
+
+    @Test
+    void reorderColumns() {
+    }
+}
